+++ web/src/declarations.d.ts
@@ -1,3 +1,11 @@
 declare const graphql: (query: TemplateStringsArray) => void;
 declare module '*.gif';
-declare module '*.json';
+declare module '*/locales.json' {
+    interface Locale {
+        [lang: string]: {
+            [key: string]: string;
+        }
+    }
+    const value: Locale;
+    export default value;
+}
