+++ web/src/pages/404.tsx
@@ -1,10 +1,11 @@
 import * as React from "react";
+import Layout from "../components/layout";
 
 const NotFoundPage = () => (
-  <div>
+  <Layout>
     <h1>NOT FOUND</h1>
     <p>You just hit a route that doesn&#39;t exist... the sadness.</p>
-  </div>
+  </Layout>
 );
 
 export default NotFoundPage;
