+++ web/src/pages/app.tsx
@@ -1,15 +1,14 @@
 import * as React from "react";
-import { FormattedDate, FormattedHTMLMessage, FormattedMessage } from "react-intl";
+import { FormattedMessage } from "react-intl";
 import ActiveApps from "../components/ActiveApps";
-import AppLink from "../components/AppLink";
 import InactiveApps from "../components/InactiveApps";
-import apps from "../data/apps";
+import Layout from "../components/layout";
 
 export default () => (
-    <div>
+    <Layout>
         <h4><FormattedMessage id="app.active" /></h4>
         <ActiveApps />
         <h4 style={{ marginTop: "2em" }}><FormattedMessage id="app.inactive" /></h4>
         <InactiveApps />
-    </div>
+    </Layout>
 );
