+++ web/src/pages/math_class.tsx
@@ -1,9 +1,10 @@
 import { withPrefix } from "gatsby-link";
 import * as React from "react";
 import { FormattedMessage } from "react-intl";
+import Layout from "../components/layout";
 
 export default () => (
-    <div>
+    <Layout>
         <h4><FormattedMessage id="math.class" /></h4>
         <ul>
             <li><a href={withPrefix(`math/lec_yogen.pdf`)}>第一余弦定理</a></li>
@@ -11,5 +12,5 @@ export default () => (
             <li><a href={withPrefix(`math/lec_circle.pdf`)}>円に内接する四角形</a></li>
             <li><a href={withPrefix(`math/lec_senkei.pdf`)}>固有ベクトル</a></li>
         </ul>
-    </div>
+    </Layout>
 );
